requirements-dev.txt
-e git+git@github.com:akaariai/fabric.git@py34#egg=fabric
pytest==2.9.1
pytest-django==2.9.1
